simulation/requirements.txt
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
humanfriendly==10.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.0
mdurl==0.1.2
networkx==3.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
prettytable==3.9.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.6.0
six==1.16.0
tzdata==2023.3
wcwidth==0.2.8
pydantic 